requirements.txt
langchain-community
streamlit
python-dotenv
langchain
langchain-core
langchain-groq
langchain-google-genai